Fix tests

+++ src/test-suites/suite_strpbrk.c
@@ -4,8 +4,8 @@ START_TEST(test_strpbrk_1) {
   char example1[256] = "\0sac";
   char example2[256] = "\0123";
 
-  ck_assert_uint_eq(s21_strpbrk(example1, example2),
-                    strpbrk(example1, example2));
+  ck_assert_uint_eq((unsigned long)s21_strpbrk(example1, example2),
+                    (unsigned long)strpbrk(example1, example2));
 }
 END_TEST
 
@@ -13,8 +13,8 @@ START_TEST(test_strpbrk_2) {
   char example1[256] = "123\0sac";
   char example2[256] = "\0123";
 
-  ck_assert_uint_eq(s21_strpbrk(example1, example2),
-                    strpbrk(example1, example2));
+  ck_assert_uint_eq((unsigned long)s21_strpbrk(example1, example2),
+                    (unsigned long)strpbrk(example1, example2));
 }
 END_TEST
 
@@ -22,8 +22,8 @@ START_TEST(test_strpbrk_3) {
   char example1[256] = "";
   char example2[256] = "\0123";
 
-  ck_assert_uint_eq(s21_strpbrk(example1, example2),
-                    strpbrk(example1, example2));
+  ck_assert_uint_eq((unsigned long)s21_strpbrk(example1, example2),
+                    (unsigned long)strpbrk(example1, example2));
 }
 END_TEST
 
@@ -31,8 +31,8 @@ START_TEST(test_strpbrk_4) {
   char example1[256] = "1234\n";
   char example2[256] = "\t";
 
-  ck_assert_uint_eq(s21_strpbrk(example1, example2),
-                    strpbrk(example1, example2));
+  ck_assert_uint_eq((unsigned long)s21_strpbrk(example1, example2),
+                    (unsigned long)strpbrk(example1, example2));
 }
 END_TEST
